End point /grupos

+++ sistema/src/main/java/com/mateus/sistema/domain/Grupo.java
@@ -20,7 +20,7 @@ public class Grupo implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String descricao;
 	
 	@OneToMany(mappedBy = "grupo")
@@ -30,16 +30,16 @@ public class Grupo implements Serializable {
 	public Grupo() {
 	}
 
-	public Grupo(Integer id, String descricao) {
+	public Grupo(Long id, String descricao) {
 		this.id = id;
 		this.descricao = descricao;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
